requirements.txt
discord.py==1.3.2
configparser==4.0.2